% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg.Obj.R
\name{avg.Obj}
\alias{avg.Obj}
\title{??}
\usage{
avg.Obj(all.A, tol.A, list.mus.hat)
}
\arguments{
\item{all.A}{A vector of all treatments that is sorted???????}

\item{tol.A}{A list of treatments that are tolerated}

\item{list.mus.hat}{A list of matrix of estimated conditional mean outcomes. Each element matrix is for one objectives,}
}
\value{
a vector indicates the average pseudo outcome of each objectives
}
\description{
??
}
