% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.frame2matrix.R
\name{data.frame2matrix}
\alias{data.frame2matrix}
\title{Converts three columns of a data.frame into a matrix -- e.g. to plot
the data via image() later on. Two of the columns form the row and
col dimensions of the matrix. The third column provides values for
the matrix.}
\usage{
data.frame2matrix(
  data,
  rowtitle,
  coltitle,
  datatitle,
  rowdecreasing = FALSE,
  coldecreasing = FALSE,
  default_value = NA
)
}
\arguments{
\item{data}{data.frame: input data}

\item{rowtitle}{string: row-dimension; name of the column in data, which distinct values should be used as row names in the output matrix}

\item{coltitle}{string: col-dimension; name of the column in data, which distinct values should be used as column names in the output matrix}

\item{datatitle}{string: name of the column in data, which values should be filled into the output matrix}

\item{rowdecreasing}{logical: should the row names be in ascending (FALSE) or in descending (TRUE) order?}

\item{coldecreasing}{logical: should the col names be in ascending (FALSE) or in descending (TRUE) order?}

\item{default_value}{numeric: default value of matrix entries if no value exists in data.frame for the entries}
}
\value{
matrix: matrix containing values of data[[datatitle]] with rownames data[[rowtitle]] and colnames data[coltitle]
}
\description{
Converts three columns of a data.frame into a matrix -- e.g. to plot
the data via image() later on. Two of the columns form the row and
col dimensions of the matrix. The third column provides values for
the matrix.
}
\author{
Daniel Neumann
}
