% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_opt.DTR.R
\name{predict_opt.DTR}
\alias{predict_opt.DTR}
\title{Predict optimal treatment using the output from DTRtree.}
\usage{
predict_opt.DTR(treeout, newdata)
}
\arguments{
\item{treeout}{An object outputted from DTRtree function.}

\item{newdata}{New data containing H (history).}
}
\description{
Predict optimal treatment using the output from DTRtree.
}
