% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_tol.DTR.R
\name{predict_tol.DTR}
\alias{predict_tol.DTR}
\title{Predict tolerant treatment using the output from MO.tol.DTRtree}
\usage{
predict_tol.DTR(treeout, newdata)
}
\arguments{
\item{treeout}{An object (tree) outputted from MO.tol.DTRtree function.}

\item{newdata}{New data containing H (history).}
}
\description{
Predict tolerant treatment using the output from MO.tol.DTRtree
}
