% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tol.A.R
\name{Tol.A}
\alias{Tol.A}
\title{The set of tolerant treatments with given tolerant rate and pseudo outcome matrix in one given stage.}
\usage{
Tol.A(A, mus.hat, tol.rate)
}
\arguments{
\item{A}{Treatment options given in this stage}

\item{mus.hat}{A matrix indicates the final counterfactual outcome according to the treatments.}

\item{tol.rate}{A scalar indicates the tolerant rate}
}
\description{
The set of tolerant treatments with given tolerant rate and pseudo outcome matrix in one given stage.
}
